Fix the https protocol check in getProtocol
The string passed into the getProtocol function is 'https:' or 'http:' (with a colon) but the check applied within does not consider this. This causes HTTP connections to be made to HTTPS servers (which results in various errors, mine was the one below).

Error I faced when using poster over a https `uploadUrl` was: `Invalid response from upload server. statusCode: 302`